+++ android/src/main/java/com/leeson/image_pickers/ImagePickersPlugin.java
@@ -62,8 +62,6 @@ public class ImagePickersPlugin implements MethodChannel.MethodCallHandler {
             Saver videoSaver = new Saver(registrar.context(),videoUrl,videoDirPath,result);
             videoSaver.download();
           }
-        }else{
-          result.notImplemented();
         }
         return false;
       }